Separate stopwatch into its own component & add vitest
