+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest",
+    "coverage": "vitest run --coverage"
   },
   "dependencies": {
     "vue": "^3.2.47"
@@ -14,6 +16,12 @@
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.1.0",
     "vite": "^4.2.0",
+<<<<<<< Updated upstream
     "vite-plugin-pwa": "^0.14.6"
+=======
+    "vite-plugin-mkcert": "^1.13.4",
+    "vite-plugin-pwa": "^0.14.6",
+    "vitest": "^0.29.7"
+>>>>>>> Stashed changes
   }
 }
